# kubeadm: introduce static defaults for unit tests