--- a/cmd/kubeadm/app/phases/kubelet/config_test.go
+++ b/cmd/kubeadm/app/phases/kubelet/config_test.go
@@ -25,8 +25,6 @@ import (
 	"k8s.io/apimachinery/pkg/util/version"
 	"k8s.io/client-go/kubernetes/fake"
 	core "k8s.io/client-go/testing"
-	kubeadmapiv1beta2 "k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm/v1beta2"
-	"k8s.io/kubernetes/cmd/kubeadm/app/constants"
 	configutil "k8s.io/kubernetes/cmd/kubeadm/app/util/config"
 )
 
@@ -51,12 +49,9 @@ func TestCreateConfigMap(t *testing.T) {
 		return true, nil, nil
 	})
 
-	clusterCfg := &kubeadmapiv1beta2.ClusterConfiguration{
-		KubernetesVersion: constants.CurrentKubernetesVersion.String(),
-	}
-	internalcfg, err := configutil.DefaultedInitConfiguration(&kubeadmapiv1beta2.InitConfiguration{}, clusterCfg)
+	internalcfg, err := configutil.DefaultedStaticInitConfiguration()
 	if err != nil {
-		t.Fatalf("unexpected failure by DefaultedInitConfiguration: %v", err)
+		t.Fatalf("unexpected failure when defaulting InitConfiguration: %v", err)
 	}
 
 	if err := CreateConfigMap(&internalcfg.ClusterConfiguration, client); err != nil {