--- a/cmd/kubeadm/app/phases/uploadconfig/uploadconfig_test.go
+++ b/cmd/kubeadm/app/phases/uploadconfig/uploadconfig_test.go
@@ -27,7 +27,6 @@ import (
 	clientsetfake "k8s.io/client-go/kubernetes/fake"
 	kubeadmapi "k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm"
 	kubeadmscheme "k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm/scheme"
-	kubeadmapiv1beta2 "k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm/v1beta2"
 	kubeadmconstants "k8s.io/kubernetes/cmd/kubeadm/app/constants"
 	configutil "k8s.io/kubernetes/cmd/kubeadm/app/util/config"
 )
@@ -50,33 +49,14 @@ func TestUploadConfiguration(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t2 *testing.T) {
-			initialcfg := &kubeadmapiv1beta2.InitConfiguration{
-				LocalAPIEndpoint: kubeadmapiv1beta2.APIEndpoint{
-					AdvertiseAddress: "1.2.3.4",
-				},
-				BootstrapTokens: []kubeadmapiv1beta2.BootstrapToken{
-					{
-						Token: &kubeadmapiv1beta2.BootstrapTokenString{
-							ID:     "abcdef",
-							Secret: "abcdef0123456789",
-						},
-					},
-				},
-				NodeRegistration: kubeadmapiv1beta2.NodeRegistrationOptions{
-					Name:      "node-foo",
-					CRISocket: "/var/run/custom-cri.sock",
-				},
-			}
-			clustercfg := &kubeadmapiv1beta2.ClusterConfiguration{
-				KubernetesVersion: kubeadmconstants.MinimumControlPlaneVersion.WithPatch(10).String(),
-			}
-			cfg, err := configutil.DefaultedInitConfiguration(initialcfg, clustercfg)
-
+			cfg, err := configutil.DefaultedStaticInitConfiguration()
 			if err != nil {
 				t2.Fatalf("UploadConfiguration() error = %v", err)
 			}
-
 			cfg.ComponentConfigs = kubeadmapi.ComponentConfigMap{}
+			cfg.ClusterConfiguration.KubernetesVersion = kubeadmconstants.MinimumControlPlaneVersion.WithPatch(10).String()
+			cfg.NodeRegistration.Name = "node-foo"
+			cfg.NodeRegistration.CRISocket = kubeadmconstants.UnknownCRISocket
 
 			status := &kubeadmapi.ClusterStatus{
 				APIEndpoints: map[string]kubeadmapi.APIEndpoint{
@@ -117,7 +97,7 @@ func TestUploadConfiguration(t *testing.T) {
 				decodedCfg.ComponentConfigs = kubeadmapi.ComponentConfigMap{}
 
 				if !reflect.DeepEqual(decodedCfg, &cfg.ClusterConfiguration) {
-					t2.Errorf("the initial and decoded ClusterConfiguration didn't match:\n%t\n===\n%t", decodedCfg.ComponentConfigs == nil, cfg.ComponentConfigs == nil)
+					t2.Errorf("the initial and decoded ClusterConfiguration didn't match:\n%#v\n===\n%#v", decodedCfg, &cfg.ClusterConfiguration)
 				}
 
 				statusData := controlPlaneCfg.Data[kubeadmconstants.ClusterStatusConfigMapKey]